Software Interrupts seem to be working OK.
Some are implemented in C, modules can be initialised and their SWI handler code called as needed.

In order to get this far, I had to implement some module handling code, and a mechanism to include RISC OS ROM modules in the build. For building without a set of ROM modules, there is a single file under Modules called DummyModule; it is just a module with a name that does nothing.

There is also an RMA per core (and a new SMA, for multi-processing aware modules' data to be shared between cores, with a simple locking mechanism).

The next step should be to reproduce the interface for DynamicAreas, for kernel memory management to appear as close as possible to the legacy kernel.

Alternatively, if possible, I'd like to hack in initialisation of the display, and have RISC OS use it.