Initialised multiple modules, but blocked at DrawFile's C implementation
Various Dynamic Areas are created and initialised, including the screen, through a hack - it will later be found and registered by a hardware-specific module.

The DrawFile module halts everything because, I suspect, it's written in C and expects a library at a certain address.

A work-around might be to copy an existing RISC OS ROM, and just replace the "HAL" portion with the new kernel. It'll be ugly, but without something concrete to show, there won't be any take-up.

I might just try using the DrawMod to write to the screen first, though.