Independently rotating cogs
This commit uses all cores to rotate Richard Hallas' RISC OS Cog, displayed using the DrawMod taken from RISC OS.

Among other things, this demonstrates implementation of:

* Module initialisation
* Provision of SWIs by modules
* Vectors (in this case DrawV, but the implementation is general)
* Some kernel services, like VDU variables
* Dynamic Areas (including the screen area, by bodge)

There is no implementation for:
* Interrupts
* Timers
* VDU/OS_Plot

On real hardware (Pi3), core 1 consistently aborts, the cache clearing doesn't work, leading to the cogs (one per core) flickering randomly, but in time that settles down and seems to work.

The Cog graphic is used with the kind permission of Richard Hallas.
The DrawMod module is part of RISC OS and licenced under the Apache 2 licence. Source is available with the rest of the operating system.