Minor changes prior to adding OS_Plot functionality
About to go away again, but real hardware failure looks like the memory pointed to by the word at address 0x20000338 + 0x124 is not mapped, or at least the MMU hasn't seen the necessary changes. (It's unpredictable which cores this will happen on.)

The above is at fc006598, there's a similar problem at fc002dc8.