First attempt at a GCC generated Relocatable Module for a HAL
I've added a couple of files to make writing a module in GCC relatively
easy, but there will no doubt be teething troubles.

It would be a good idea to include optional checks on vector calls to ensure
that the registers that are supposed to remain untouched actually are.

At the moment, the module still doesn't access any hardware, just the frame
buffer, but it does claim the WrchV vector and display four independent
"consoles" (output only) of 40 rows of 60 characters. Control characters
appear in red.

The next step is to request the kernel grant access to device memory wanted
by the HAL (the mailboxes, for example), and also to establish a regular
timer interrupt.