Working display, but only because of undue knowledge of the pi in the kernel
I've been trying to get the HAL to initialise the display instead of the
kernel, which is tricky when it doesn't work on real hardware and you can't
see what's going on. It works fine in qemu, of course.

Strangely, if the frame buffer has been initialised once before the HAL's
attempt succeeds. The HAL is communicating with the video core via mailboxes
correctly, and the tag interface works, but only if the kernel has already
used it.

Maybe I should just try to repeat the request twice in the HAL?