HAL module is now responsible for initialising the screen
All four cores update their "console" output on a 1080p screen, until there's
a data abort during the initialisation of BufferManager.

I'm travelling for a week, so an incomplete commit. But it works in QEMU, and
at least one physical raspberry pi! I should probably upload the occasional
ROM image.

There is still an issue with initialising the HAL before core 0 has reported
available RAM (and, of course, the HAL module has to report the full amount
of RAM in the device available to the ARM).

The virtual screen size is slightly taller than the visible size, so that
the bitmap completely fills an 8MiB section of memory without any chance of
the ARM interfering with the GPU.