Enough time spent down the WindowManager rabbit hole
It looks like the Window Manager does too much weird stuff (why does it
need to know the address of the ROM, for example) to be left unchanged.

The next approach will be to add TaskSlots and threads. Threads will have
thread-local storage at the top of their allocated stack memory.