Vector calling fixed
It took me a while to recognise the importance of this section of
documentation:

 "If your routine passes the call on, you can deliberately alter some of
 the registers values to change the effect of the call, however, you must
 arrange for control to return to your routine again to restore to those
 that the original routine would have returned. It should then return
 control back to the calling program."

The only way I can see this working is if the code:
    Stores the intercept return point (and its private word?)
    Replaces it with an address in its own code
    Returns with mov pc, lr (allowing other handlers to execute)
AND the final action of every default vector handler is pop {pc}.

Anyway, that now works, afaict, and the next problem is
  Bad heap descriptor: FileSwitch

The build script and kernel.c have been updated to allow builds that
will only use a single core for most of the boot process, blocking all
but core 0.

./build -DSINGLE_CORE

This is useful for debugging. The HAL also sends every byte passed to
WrchV on core 0 to the UART, which can aid logging.

qemu-system-arm -bios rom.bin -machine raspi2b -nographic | less