Trying to get HLine to work properly using legacy variables
I've implemented a simple vdu stream handler to call OS_Plot when a VDU 25
stream is noticed.

If the legacy vector pointer for WrchV is < 0xfc000000, OS_Plot will create a
stream of VDU 25... characters. Setting it to 0xfc000000 avoids that.

Trying to display a triangle caused plenty of calls to HLine, but nothing
happened on the screen.

To minimise the code to be run, I've started debugging a direct call to
ExportedHLine from just after the HAL module has initialised (mapping the
screen).