Assorted changes that will only work with a custom ROM image.
The build script now expects at least one argument, the RiscOS directory of
the custom ROM image (I export mine from a Pi system using !Moonfish).

The output file is now "latest.bin".

Any other arguments are passed to gcc while building the kernel and modules.

Particularly useful extra C flags are:

-DSINGLE_CORE
-DDEBUG__SHOW_FONT_PATHS
-DDEBUG__SHOW_SERVICE_CALLS
-DDEBUG__SHOW_SYSTEM_VARIABLE= <1..7>
-DDEBUG__FREE_POOL
-DDEBUG__SHOW_COMMANDS
-DDEBUG__SHOW_FONT_PATHS
-DDEBUG__SHOW_GSTRANS
-DDEBUG__SHOW_HLINES
-DDEBUG__SHOW_MODULE_COMMANDS_ON_INIT
-DDEBUG__SHOW_MODULE_INIT
-DDEBUG__SHOW_MODULE_LOOKUPS
-DDEBUG__SHOW_OS_BYTE
-DDEBUG__SHOW_RESOURCE_FILES
-DDEBUG__SHOW_SERVICE_CALLS
-DDEBUG__SHOW_SYSTEM_VARIABLE
-DDEBUG__SHOW_TRANSIENT_CALLBACKS
-DDEBUG__SHOW_VDU_VARS
-DDEBUG__SHOW_VECTOR_CALLS
-DDEBUG__WATCH_DYNAMIC_AREAS

Defining too many of the above can lead to the kernel exceeding 64K, which
isn't catered for. It would require the ROM build to move the kernel up in
memory (making the space for the HAL bigger), and a change to the dd command
in the script to have read !!!!OSBase or KernelBase from Kernel_gpa.