Various improvements, added bugs, and integrating the RISC OS build tree
The build script now expects a (link to a) directory containing a
standard build tree in the main directory called RiscOS. I have a symbolic
link to a NFS mount point that resides on a Pi running the Moonfish NFS
server. Maybe hosting the tree on the Linux box and accessing it with
Sunfish from the Pi would be better.

A script in the tools directory usually manages to find the RO code
associated with a given ROM address. (Modules written in C, not so much.)