Reworked MMU code working nicely
Hopefully, the approach is clear and relatively simple (if incomplete),
unmapped virtual address sections and pages have the address of a
function in their translation table entries. That function will try
to ensure that memory is made available to the code.

Permission faults will be dealt with earlier and differently to the
translation faults due to missing memory. That's still a TODO.

I made a slight modification to Desktop/Wimp/s/Wimp01 getromsprites,
jumping straight from just after the sub sp, sp, #4 to just before
  ADR     R1,spritesfname  ; R1 -> original filename

The intervening code never got that far. It helps a bit, but not much.

Next step: a better implementation of OS_PipeOp pipes than can block
a Task while waiting for space or data.

Perhaps the HAL will start a Wimp program with a window for each
core's debug output... That really needs fonts to be working, though.