Re-written debug output, fixed some ThreadOp functions
The debug output from the kernel now goes via a pipe, rather than
WrchV. It's much faster, less likely to mess up the display, and
can, in future, be displayed in Wimp windows, not just on the screen.

Timer interrupts are also working, slowed down for testing, so other
interrupts should be possible. Sleep works.

Interrupted tasks that are in SVC are resumed in reverse order, to
maintain the integrity of the SVC stack. I'm hopeful the SVC stack
will never need blindly resetting.