Fonts finally close to working; HLine is used for sprites, not just the screen
And ScreenStart is not necessarily anything to do with the screen, either.

A lot of details still need hammering out, like where is the colour information
supposed to come from (I'm sure I'll eventually find out by reading the legacy
source, but still...)?

Next problem: deadlock on startup with multiple cores (not just on real hardware).