Many changes
Re-implemented GSTrans (and GSInit/GSRead in terms of it), still using legacy Set/ReadVarVal.

Timers and interrupts improved (hopefully fixed) so at least Sleep works OK. (Timers are still relatively slow to allow for running in QEMU.)

Added some file support to newlib (barely started, nowhere near complete).

About to change pipes to be owned by slots, not tasks, so that it's easier to provide services. I think.

I should merge the branch with the main branch and make a new one for the pipes changes....