Adding SWIs to MTWimp module
The idea is to make the Wimp more responsive without having to modify
the legacy code.

The MTWimp module starts a Wimp Task, polls, then holds the door open
until an event that the Wimp can react to occurs. At that point, all
Wimp tasks think they're in a Wimp_Poll(Idle) call.

It's also a prototype for a Task that can perform services for other
tasks without locking up the Legacy kernel code.

A requesting task will call a SWI, the SWI will safely store the request
and the Task making the request in the module's workspace, wake the
module's server Task and wait until woken again.

We need a mechanism to say a SWI is legacy-free, i.e. it doesn't need
the full SVC stack (currently per-core, but that might change if it
continues to be helpful to only allow one Task entry to legacy code
at a time).

Either a module flag, or a SWI bit like the X bit?

Only the barest bones have been implemented and not tested.