Many changes, to do with moving the implementation of GSTrans etc. into usr32
See the files swis/varvals* to see how a service thread can run code in
another slot, and use pipes to access that slot's memory.

The next change is to smooth out the Kernel_physical_address etc. routines
so that all memory regions are covered, and described in just one place.